ocr/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
google-cloud-vision==3.7.2
pdf2image==1.17.0
Pillow==10.4.0
